Document xdebug_break() behavior.
xdebug_break() is helpful when the debugger won't stop at a breakpoint.
The causes of this vary; for example some complex auto class loading may
be confusing xdebug. Using xdebug_break() is a last resort, but at least
it achieves the desired effect of breaking at a line of code when
nothing else works.