Hack to allow the vim-tmux-navigator plugin key bindings to interupt.
This allows the <c-h>, <c-j>, <c-k> and <c-l> binding provided by the
vim-tmux-navigator plugin to be used to switch to a tmux pane while
vdebug waits for a connection to the debugger.

This is useful when debugging cli scripts and you need to change to a
differnt tmux pane so as to execute your script.