Added feature of context_get in any stack depth
Now Vdebug can query any stack depth, and the corresponding variables
and values will appear in the watch window. Also, the stack depth file
and line number will also appear correctly.

A visual sign (#>) was introduced to mark clearly the source and stack
windows for the current stack depth and file_and_line. When the
debugging resumes, these stack signs are removed from the ui.

Also updated the docs with a short description of these features.