Property expansion enhancement in Watch window - all children
Now a property expansion in Watch window list all children from that
property. The previous behavior was listing only <pagesize> elements
(i.e. based on the pagesize attribute of parent element).

As before, to invoke a property expansion either double-click it or
place the cursor over it and hit Enter.