Tag known failure in integration test
The tag is used to mark the scenario as pending before it is run.  This
allows the travis builds to pass and makes it easier to spot other
failures.  Sadly this does never run the test so we have no indication
when it is fixed.  The tag has to be removed manually in order to check
if the test was fixed.