make sure temp keymap is read in correct encoding
Apperantly mkexrc stores a file with cp1250 encoding, when re-reading
the file and trying to use it as utf-8 you get errors. So no we read the
file with the cp1250 encoding and transfer the data back to utf-8. This
avoids a supressed crash when a debugger connects to vdebug and there
are utf-8 chars in the configuration that get misinterpreted because the
faulty exported data.

fixes #337

Signed-off-by: BlackEagle <ike.devolder@gmail.com>