open the exrc tmp file with ignore errors
When trying to store and restore the keymapping, the encoding might
differ, initially we found that most of the time the mkexrc command
would store in cp1250 encoding. It seems that depending on the vimrc
that command uses another encoding. Now we just open de tmpfile and
ignore encoding errors.

fixes #330

Signed-off-by: BlackEagle <ike.devolder@gmail.com>