counter mkexrc fileencoding behaviour
mkexrc writes the exrc file with an encoding detected somewhere through
your vimrc file, ...  With a properly encoded vimrc and such you get a
nice utf-8 file, with some vimrc's there is some mixing which causes the
mkexrc command to write the exrc file with latin1 or cp1250 or ...

Now the reading of the temporary exrc file is done binary and the
strings are converted back to utf-8 using the replace strategy. This
should make sure we don't get Unicode errors while reading the file
searching for maps that need replacing for proper functioning of vdebug.

fixes #350

Signed-off-by: BlackEagle <ike.devolder@gmail.com>