use double quote around property in property_get
Some debuggers don't like you just sending obj['subobj']. The possible
response you get there is that the property is not found. Refering to
the dbgp documentation we an also send the property_get command with the
property between double qoutes, but that implicates we must escape
backslashes and double quotes.

Fixes #312

Signed-off-by: BlackEagle <ike.devolder@gmail.com>