only neovim has vim.funcs
when checking if we can use sign priority we must make sure to check if
vim.funcs exists before using it because its only available in neovim.
You don't see this issue poping up if your version of vim is recent
enough since it will not contine into the or.

Fixes #474

Signed-off-by: BlackEagle <ike.devolder@gmail.com>