add default feature set of some features
If we look at https://xdebug.org/docs/dbgp#options-and-configuration in
the feature names we see any dbgp debugger must provide some feature set
by default. Even if that feature is not actually supported.

So from now on we will set multiple_sessions = 0 because Vdebug does not
support this at the moment and we will try to set extended_properties =
1; if the used debugger does not support it we will silently continue
without the feature.

fixes #369

Signed-off-by: BlackEagle <ike.devolder@gmail.com>